refactor(quality): move workspace_context.py into workspace/ package (P4)
Moves the orphan top-level module into the existing workspace/ package as
workspace/context.py. workspace_context.py becomes a thin backward-compat
shim for any unupdated callers.

All 13 src/ callers and 6 test files updated to import from the canonical
location. Patch targets in test_review_validation_unit.py updated from
specify_cli.workspace_context.load_context to specify_cli.workspace.context.load_context
so monkeypatches land in the module where the function actually lives.
