+++ src/specify_cli/workspace/__init__.py
@@ -14,9 +14,49 @@ from .root_resolver import (
     canonicalize_feature_dir,
     resolve_canonical_root,
 )
+from .context import (
+    NormalizedWorkPackage,
+    ResolvedWorkspace,
+    WorkspaceContext,
+    build_feature_context_index,
+    build_normalized_wp_index,
+    clear_workspace_resolution_caches,
+    cleanup_orphaned_contexts,
+    delete_context,
+    find_context_for_wp,
+    find_orphaned_contexts,
+    get_context_path,
+    get_normalized_wp,
+    get_workspaces_dir,
+    list_contexts,
+    load_context,
+    resolve_feature_worktree,
+    resolve_workspace_for_wp,
+    save_context,
+)
 
 __all__ = [
+    # root_resolver
     "WorkspaceRootNotFound",
     "canonicalize_feature_dir",
     "resolve_canonical_root",
+    # context
+    "NormalizedWorkPackage",
+    "ResolvedWorkspace",
+    "WorkspaceContext",
+    "build_feature_context_index",
+    "build_normalized_wp_index",
+    "clear_workspace_resolution_caches",
+    "cleanup_orphaned_contexts",
+    "delete_context",
+    "find_context_for_wp",
+    "find_orphaned_contexts",
+    "get_context_path",
+    "get_normalized_wp",
+    "get_workspaces_dir",
+    "list_contexts",
+    "load_context",
+    "resolve_feature_worktree",
+    "resolve_workspace_for_wp",
+    "save_context",
 ]
